Lowercase rabbitmq nodes
If nodenames use uppercase then there is a mismatch because
rabbitmq uses host's hostname locally and hostname is
lowercased by cloud-init.

This is not a problem by default because controller node names in
default heat templates are lowercased. But this issue can be observed
if an older/different template is used.

Change-Id: I71ed191c83b146b89f140e231920bb3621983bbe