Use keepalived notify script if virtual_interfaces are defined
Notify script which sets explicit routing for VIP should be used
no matter if virtual_interfaces is defined or not (it was set by
mistake for no virtual_interfaces configurations).

Note that notify script is not invoked when keepalived process itself
terminates. In such case the explicit route for VIP is not removed
and traffic for VIP is routed to localhost - this could be improved
in future if we find out that keepalived itself is unstable (I didn't
hit issues with keepalived itself when testing it).

Related-To: 1376200
Change-Id: I1ebb62a4d7b52bf558d904ac01040b75d5d00d4e