Swift proxy memcache authtoken additions
Now that memcache is enabled on Swift proxies, we want to make use
of it for caching of Keystone authtokens.  There are two parts to
this.  First of all, when Swift memcache is enabled - as confirmed
by the existence of swift.proxy-memcache metadata - we want to
use this cache for authtokens by setting
   cache = swift.cache
in the [filter:authtoken] stanza.  Secondly, many Swift deployments
will want to encrypt these cached tokens for security reasons.  To
enable this, we set the swift.encrypt-authtoken metadata to
something other than null.  Furthermore, we need to provide an
encryption key which will need to be unique to each deployment.
We use the value of swift.hash for this key.

Change-Id: I969dc6753224aed7165ba264d3fddf6f6fa1542b