Don't rebuild swift rings if already exist
The swift element os-refresh-config builds the swift account, container
and object rings and overwrites previous versions. The swift element
is only used to build the initial rings, it is not intended to be used
to manage/extend the ring. The user must manage the ring outside of
tripleo. This change prevents the swift element from rebuilding the
the rings if they already exist.

Change-Id: I3abe515432feeb7820388dc013244b2fa9fb3bc6