Use nagios-common instead of nagios-plugins-all on RHEL.
nagios-plugins-basic was mapped to nagios-plugins-all on RHEL, however
nagios-plugins-all is actually not shipped with RHEL. The reason it
seems to be installed via the snmpd element is so that the nagios
plugins directory exists.  This can be accomplished on RHEL by
installing nagios-common instead.

Migrates all of the element over to pkg-map as well for the other
installed packages.

Change-Id: I8773a93dcd8b348b501d8303ddd211b3edf1e2b9