NTP fail to sync if time sync is not performed
Time sensitive services may not be able to startup and be synchronized,
specifically when deployed on multiple nodes, when the hardware clock
of the booting system is not explicitly synchronized before attempting
to start the services. With this, NTP itself may not be able to
synchronize and start operating if the clock is skewed too much from
the clock source(s), or if one clock source is out of sync itself.

This change move the ntp stand-up earlier in setup so that time of any
system is controlled befor time sensitive services startup. For example
the database on controller nodes. The change also set the clock to the
time of the given ntp server forcefully and updates the Hardware clock.
It also fixes the invalid ntp.conf file.

Closes-Bug: 1408379
Relates-To: I56a03dc0a899a8c515f2a05d678d7e80e9b7b93c
Change-Id: Ib9931b84925d9ceb32f18e9adc5be64402fbf61e