Prevent unclustered RabbitMQ with 0 nodes
It can happen that RabbitMQ attempts to start before Heat has delivered
rabbit.nodes. It comes through as a blank value, IE with 0 nodes in it,
causing the Rabbit post-configure script to think we want a
non-clustered Rabbit config. This is not limited to one node - all nodes
will attempt to start non-clustered.

They can then start receiving messages, which makes joining a cluster
later difficult. To avoid this situation, first of all only allow the
bootstrap node to come up non-clustered. If you want a non-clustered
config you must have a single node control plane already so this is
fine.

Do not configure at all if we don't know who the bootstrap node is. Nor
should even the bootstrap node configure as non-clustered if we can see
other nodes are clustered.

Change-Id: Ib2ec261c84143640a5ccd1abbc99be1c8020c543