Delorean: going back in time fails,package caching
This fixes an issue where setting your DELOREAN_REPO back
in time (to an earlier repo) doesn't work. This is
because the Yum metadata cache for the more recent
repo is already stored in your
.cache/image-create/yum/x86_64/20/delorean and it won't
get updated unless it is manually cleaned.

This patch adds a simple 'yum metadata clean' command
to the delorean-repo element so that we can always
ensure the metadata file we want gets downloaded.

Change-Id: Ie714d7298c9691ec156f4013ccd3edb2cc0912d1