Revert "Add support for pxe_ilo driver"
This reverts commit 26df595990576c68b2343c2758808cb02de0d642.

This commit breaks distros that do not have proliantutils
packages. This is because Ironic is written to be fail
fast with regards to driver dependencies. Any driver
that is listed in the config file must have its dependencies
present.

We should not require all distros to carry these packages.
Rather we should implement a pluggable sub-element that
can be used only on distributions which support this
driver.

Change-Id: I8ed74d094b1c5555c1ac9332caabadbfd02cabdf
Closes-bug: #1411809