Ceilometer API fails: bad rabbit_use_ssl setting
This patch updates the ceilometer.conf file to avoid an issue
where rabbit_use_ssl got set to <blank> if no Heat metadata
for this setting is provided. Some of our heat templates
(the undercloud for example) do not currently provide this
information and as such Ceilometer API fails to startup
correctly in those deployments.

Change-Id: I0a7dcfb54a0f113e0a3e59e3ccedfe0f485bf12d
Closes-bug: #1416499