Default to latest rdo-release
Currently one needs to explicitly specifies the $RDO_RELEASE one wants
to retrive. In order to alleviate the number of environment variables to
export from the user side it sounds reasonable to default to the latest
available version. The same process is used for the Fedora release to
install in the fedora element[1]

[1]
https://github.com/openstack/diskimage-builder/blob/master/elements/fedora/environment.d/10-fedora-distro-name.bash#L2

Change-Id: I5d983ecc03bd157cdbc707b70975bdd8d8c59ced