os-net-config: add ping_metadata_ip function
Updates os-net-config to use --detailed-exit-codes so that when
a network change has been detected we can perform a network validation
via a new ping_metadata_ip function. This function performs a simple ping
test back to the IP address of the Heat API which should always be
available after any network configuration change.

The motivation for this change is to allow network configuration
changes time to settle. When making changes to the ctlplane (which
is typically used to contact the undercloud Heat API for subsequent
configuration notifications and updates) it is sometimes desirable to
wait a few seconds before proceeding. This avoids a race where the
ctlplane was switch to a static IP address and the very next
Heat notification would sometimes fail.

Depends-On: I8f22fa15335d1276f4e444a6454a24ff486e1495

Change-Id: I257e1cba6dee16f73f75512d1284e1e3b9d4c831