Don't set default install type for os-cloud-config to pip
When this element was first added, pip was set as the default install
type.  This is not ideal since it means that $DIB_DEFAULT_INSTALL_TYPE
is not honored when using this element, you'd have to explicitly know
that you also need to set $DIB_INSTALLTYPE_os_cloud_config.

This patch removes the settings of $DIB_INSTALLTYPE_os_cloud_config so
that $DIB_DEFAULT_INSTALL_TYPE is honored.

Change-Id: I8f47b495218a72f5a78d68b7c2610116a5b4984c