Remove mysql-dev dependency from os-svc-install
Looking through the image build logs, I noticed we were pulling in
the mysql-devel packages for some reason.  It turns out this is a
historical artifact of the old TripleO source installs using
os-svc-install.  It was universally installing mysql-python on
every image where it was used because that package wasn't in the
requirements.txt for the OpenStack projects.

This was wrong in the first place, and now that we aren't supporting
that installation method anymore it doesn't make sense to keep
installing nine development packages that we'll never use on our
images.

Note that the only thing we use from os-svc-install now appears to
be os-svc-enable, so this should have no effect on the current
TripleO flow, except slightly smaller/faster image builds.

Change-Id: Iae3e6ae3f57de57639c4259361e82c8150fe2f0e