Use package-installs for undercloud elements
Ideally all elements would be using package-installs so we only
call yum once per stage, but since a lot of these elements are no
longer used this change focuses on just the elements doing one-off
package installs that I found in the undercloud install logs.

A follow-up change would be to do the same kind of audit for the
overcloud image build and fix up all of those elements.

Change-Id: Ie204c5d0aa0f48d3d93d834b51a644d8562cd2fd