Purge chrony package in NTP element
If NTP is installed and used, make sure Chrony is not installed because
they would conflict each others.
We have seen some cases where NTP was started after Chrony but failed
because of port conflict.

Since Chrony is installed by default in EL7 images, we want to avoid
this conflict.

Change-Id: I186152786a27ca5bee1b288a76c2e340971596a4