Dummy patch to build a new rpm in RDO
The version of openstack-tripleo-image-elements in RDO for stable
branches is too old and it might be one of the reasons why TripleO CI is
failing to deploy TripleO on Newton.

See: https://bugs.launchpad.net/tripleo/+bug/1690373

This patch is doing nothing important but by merging it, we'll build a
new package in RDO for master and all branches and hopefully get a new
build that will help us to resolve the issue.

Change-Id: Ie6c21aa8d071e791eb0df6db6d700074d3141cf0
Related-Bug: #1690373