Update ironic-agent-multipath element
It finally was tested on real environments, and it was not
working properly. The generated image is not build with
dracut, so the regenerate we did there was not used, and the
modules were not enabled.
Adding the modules on the right directory, to be loaded at
boot time.
Also removing the 98-regenerate-initramfs patch, because it
was regenerating an image that is never used, so it was a
no-op

Change-Id: I87527a4ac2111f1e572e3c779e11fdf20fc88bc8
(cherry picked from commit 22a5e4e50f2bf2a71128614218ed208ee8f6f5c2)