Increase the image size to 23G
The recent tests showed that we need more space for the /var
partition, and that we need to increase the final size of the
image to 23G to fit the needs. And we need to rebalance the
percentage used by volumes to have similar sizes.

Change-Id: Icf25fc15bd71a50e68038012683e3e56aaa72443
Depends-On: I3885a5c359b56c515dbb9286ded1dee062cdcffc
(cherry picked from commit e86b30a73f6a80da137ad0a5d26b96c263d33bcd)