Add a dependency on block-device-mbr element
When specifying the layout for the hardened image, we need
also to specify the device type, that in our case will be
mbr. So add this as a dependency of overcloud-secure to
properly set the type.

Change-Id: Ic0ad151cc12723e437a1123f90c6663198b9564d
(cherry picked from commit 950a63f77c8c9da9ce6c7fccb80fabd980a4e4bc)