Clear net.ifnames definition
With RHEL8, net.ifnames is set to 0 in the grub config. This is a change
in the default image which can have user impacts. We want to clear it to
go back to the previously defined behavior in RHEL7.

Change-Id: I621066ea5e0e934668ff4c91e991a95d18e019f7
Closes-Bug: #1841441
(cherry picked from commit 8da8d9a51b287d520c465fae254262fc4286e92e)