Cleanup stale interface if exists
There is a bug in the CentOS 8 image where the ens3 interface file
exists. We should clean that up if it exists to prevent issues when
booting. We manage the interfaces later with os-net-config so we don't
want them to exist in the image.

Change-Id: I95d851f194a524caad639188e7df8f041fa2a248
Closes-Bug: #1866202
(cherry picked from commit d641e019b90f9318a38d67b3910cb4e4b1c09ec0)