Unmask and enable tmp.mount for overcloud-partition
In recent centos 8 images tmp.mount is masked, which causes issues
with all defined LVM mounts (/ is readonly, other mounts fail).

This change unmasks and enables tmp.mounted so the existing LVM /tmp
mounts correctly. It also makes fsck-passno compliant with the
documentation[1] so that / has value 1 and all other mounts are set to
2.

Blueprint: whole-disk-default
[1] https://www.man7.org/linux/man-pages/man5/fstab.5.html

Change-Id: Ib517e73845a7698f873a4f3207ffa9c54cdbcbef