Clean up ifcfg-eth0
If ifcfg-eth0 is defined but does not exist, this can cause the nodes to
fail to get an IP address on initial provisioning. This patch removes it
when the image is built to ensure we don't end up with a bad interface
network script.

Change-Id: I3b8cb5f8c413c109da1228878db9cff4d439273b
Closes-Bug: #1931495
(cherry picked from commit f90a830fbac3b9ac1e501a768eeb7454e4d540f8)