Shrink whole disk image from 40G to 6G
Now that images are converted to raw on the undercloud we need to
consider the implications of creating a 40G whole-disk image which is
mostly empty.

The raw image file on the undercloud is sparse, so storage space is
not consumed. However transferring from the undercloud to deploying
nodes will transfer all 40G of mostly zeros, and ironic-python-agent
will also write all those zeros to disk during deployment. This makes
deployment unnecessarily slow and will cause network scalabilty issues
for large overcloud deployments.

This change reduces the size of the physical root partition to 6GB,
and makes the logical volumes big enough for space to not run out
during image building. The growvols utility will be used post-deploy
to grow all of the volumes (not just /var).

The growvols playbook now has the following default to ensure all
partitions are grown to an appropriate size:
/=8GB /tmp=1GB /var/log=10GB /var/log/audit=2GB /home=1GB /var=100%

Change-Id: I34b57958ba84714c54ed8148600e7fd0860e3ba8
Blueprint: whole-disk-default