Add small /srv volume for optional swift data
Most tripleo node roles store their state in /var with the exception
of the ObjectStorage role which stores its state in /srv.

This change adds a very small /srv volume which will generally remain
empty.  A documentation change will be proposed which shows how to set
custom growvols values to grow /srv in the rare case that the
ObjectStorage role is used.

Change-Id: Ide8776ffdb62bff46c3df4773fc7b01e9710373f
Blueprint: whole-disk-default