Add element reset-bls-entries
This is an element which will run a script on first boot to rename
`/boot/loader/entries` filenames to match the new value of
`/etc/machine-id`. This is required for `grub2-mkconfig` calls to
incorporate existing values in the boot loader entry files.

It only runs when first-boot criteria is met, such as when
`uninitialized` is written to `/etc/machine-id`

rhbz#2073855

Change-Id: I354aeabed6d2993ae600b86381a7b625ca7f89dc
Depends-On: https://review.opendev.org/c/openstack/diskimage-builder/+/837251