Switch to MiB for all partition/volume sizes
This change switches to MiB as the unit for all partitions and LVM
volumes. It includes comments stating overheads so any future changes
can be easily calculated.

All LVM MiB values are multiples of 4MiB, which is the size of an LVM
extent.

This change is in preparation for thin provisioning, where specifying
volume size by percentage no longer makes sense.

Change-Id: I768bb6def4fcea96a59357b4117d92efa3b9613b
Depends-On: https://review.opendev.org/c/openstack/diskimage-builder/+/839829