Switch to LVM thin provisioning
This change switches all LVM volumes being backed by a single thin
provisioning[1] pool.

This will allow day-2 operations to use spare physical space for
things such as:
- backups during a major upgrade on an ephemeral volume
- extending /srv or /var due to inaccurate storage capacity planning

This change enables the fstrim.timer unit on a daily schedule so that
the thin provisioning layer is notified of freed blocks.

[1] https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/9/html/configuring_and_managing_logical_volumes/creating-and-managing-thinly-provisioned-volumes_configuring-and-managing-logical-volumes

Change-Id: Id1c7e627b67b930347eed040176a9737cabc87d0
Depends-On: https://review.opendev.org/c/openstack/diskimage-builder/+/840144